This removes anchor tags that were rendering icorrectly and fixes the incoming link to point to the actual ID generaged
It seems on docs.microsoft.com anchor tags (`<a name="...">`) only render correctly (i.e. invisible anchors) if they following title does not have its own ID. Otherwise the anchor is rendered as visible text and become non-functional.

We had two anchor tags that had this problem. One of them had an incoming link which I have fixed to the the actual id on the following title.

There are two more anchor tags on the installation documeation which seem to still be working correctly so I am not touching.