Removing more anchors that are not rendering correctly
It seems on docs.microsoft.com anchor tags (<a name="...">) only render correctly (i.e. invisible anchors) if the title that follows them does not have its own id. Otherwise the anchor is rendered as visible text and becomes non-functional.
We had two anchor tags that had this problem. One of them had an incoming link from our own documentation, which I have fixed to the the actual id on the title after it.
There are two more anchor tags on the installation documentation which seem to still be working correctly so I am not touching them.