Experiment: Moving roadmap and other wiki content to docs
Fixes https://github.com/aspnet/EntityFrameworkCore/issues/9615.

The current roadmap in the repo's wiki is out of date.

Wanted to see how moving the roadmap to become part of our documentation (alongside the release history) feels and look like.

If we go ahead with this, I will replace the existing pages in the docs with links to the documentation.

Note that here I also moved an explanation about the backlog (that was part of the roadmap wiki page) and the release planning process content here.

I have also removed "critical" and "high priority" lists that as explained in #9615, have become the source of confusion.

Open to suggestions.