Remove null propagation from Load calls
Both the extension methods are specifically designed to be called on a null object without failing and will return the contents of the backing field regardless of whether the lazy-loading object (ILazyLoader or delegate) has been set. Null propagation breaks this because it causes null to always be returned from the getter if the lazy-loading object is not set.