C42263: Broken hyperlink and misplaced asterisks (#1013)
* C42263: Broken hyperlink

Localization team has reported source content issue that causes issue for localization. 
Description: The link doesn't seem to work anyway. So either it is needed to update the link or place it in the box. Also there are extra asterisks, which werent reported as a bug (LOC and ENU is matching).
Please review the comment tag added and reply with explanation if fix is needed or not. If you make related fix in another PR then share your PR number so we can confirm and close this PR.
Many thanks in advance.

Fixed these and other similar issues.