Update data-seeding.md
I propose the removal of "the" from the following paragraph:
"Alternatively, you can use `context.Database.EnsureCreated()` to create a new database containing the seed data, for example for a test database or when using the in-memory provider or any non-relation database. Note that if the database already exists, `EnsureCreated()` will neither update the schema nor the seed data in the database. For relational databases you shouldn't call `EnsureCreated()` if you plan to use Migrations."