Modified the VIEW
In the code snippet for the database object VIEW, it uses the "Name" field without the alias "b" which all other column names have. 
Just to be consistent, I propose adding the alias i.e. b.Name as opposed to Name.
Lastly, I have capitalized the "from" to FROM for readability and consistency as it is a reserved word like JOIN, GROUP BY etc.