Made the property DateCreated in class BlogDetails non-nullable.
It appears the intent in the ComplexType code example was to make the DateCreated property non-nullable.  This is confirmed by this excerpt in the document: "Another interesting note is that although the DateCreated property was defined as a non-nullable DateTime in the class, the relevant database field is nullable. You must use the Required annotation if you wish to affect the database schema."