Adding preview 4 features
We are light on features for preview 4.

I have not included any of the collation support since the last part of that is in preview 5. (Yes, we should have pulled it in to 4.)

Also, I would like to include examples of better translation exception messages, but I'm having trouble finding them.