Fix the CreatePoint() example (#2404)
The Point.CreatePoint() method has two variants:
        public Point CreatePoint(CoordinateSequence coordinates);
        public Point CreatePoint(Coordinate coordinate); 
It does not support a variant that takes two decimals as described in this example. Visual Studio 2019 marks the current version of line 90 as an error.
Rather those decimal values need to be wrapped in a Coordinate object that can be constructed from those values.
I have updated line 90 to reflect this change.