add EFCore.SoftDelete (#2500)
* add EFCore.SoftDelete

An implementation for soft deleting entities for EntityFramework Core.

* add efcore.softdelere extension

an implementation for soft deletes in efcore 3.1 nad later

* Add EFCore.SoftDelete

An implementation for soft deleting entities for EntityFramework Core.

* Ass EFCore.SoftDelete

An implementation for soft deleting entities for EntityFramework Core.

* Add EFCore.SoftDelete extension

An implementation for soft deleting entities for EntityFramework Core.

* Add EFCore.SoftDelete

An implementation for soft deleting entities. For EF Core 3.

* Add EFCore.SoftDelete extension

An implementation for soft deleting entities. For EF Core 3.