Remove ms.assetid
The purpose of ms.assetid was to assign a unique ID to each doc page so when we need to look up for the same page in old publishing system (in case something did not migrate correctly), we can correlate and find out missing information from old system. (Generally missing blocks or code snippets)

They are rarely used. In case if need to use them, we can refer to this commit to find the assetid.