Docs for simple logging (#2740)
Fixes #2085

Note that this is the first part of several new pages/updates for logging and interception. For now, I'm linking to the new page from the old top-level logging. As more of this gets done that will gradually be replaced. I didn't add anything to the table of contents yet so as not to make this too confusing.