ForSqliteHasSrid() method is Obsolete
In the "Configuring SRID" section of the SQLite Spatial documentation there is an example that uses the "ForSqliteHasSrid()" method. According to the [.NET API Browser](https://docs.microsoft.com/en-us/dotnet/api/microsoft.entityframeworkcore.sqlitepropertybuilderextensions.forsqlitehassrid?view=efcore-3.1) this method is obsolete and has been replaced with the "HasSrid()" method in EF Core 5.0. This change replaces the references to the obsolete method with references to the current method.