Mention about --startup-project param
Is important to mention that should be inside data folder and need to use --startup-project to run migrations. I saw a lot of questions opened in stack overflow about it. Actually i spend a lot of time trying to discover a solution until tried to do like i mentioned before.