Fix links from Breaking changes to What's new
Links which did not use xref were broken (lines 136 and 159). Line 103 worked but I wanted to unify it with the rest. There are two options:

1. use xref - longer path, doesnt require .md suffix
2. use just file name - required .md suffix to work in github

I went with option n.2. Not adding the .md suffix would probably work on docs.microsoft.com website but doesn't work when browsing on github.