Fixed GetEnumerator() method on mock
The Mock<IQueryable> needs to be configured to call the underlying collection's `GetEnumerator()` method every time it is invoked. As it is currently written, the method is only invoked once (during setup) to get a single value and that value is returned each time `GetEnumerator()` is called. This causes issues if the underlying data collection is changed by adding and/or removing items. Doing so results in the enumerator becoming stale and causes Invalid Operation Exceptions if you try to iterate through the IQueryable or call extension methods like  `ToList()` or `ToArray()`