Add EntityFrameworkCore.AuditInterceptor to EF Core extensions list (#4961)
This commit adds EntityFrameworkCore.AuditInterceptor to the list of available EF Core extensions.

EntityFrameworkCore.AuditInterceptor is a library that provides seamless auditing capabilities for Entity Framework Core applications. 
It automatically tracks entity changes, recording who made the changes and when they occurred through a simple implementation of the IAuditable interface.