Highlight that migrating from EF6->EF Core requires being off .NET Framework
I have seen a lot of confusion around when to perform the EF6->EF Core migration while migrating applications from framework to core. This adds a warning that EF6 is supported on all supported targets, while EF Core is only on .NET Core/.NET.

It also removes a comment around ApiPort that is no longer valid since that tool has been deprecated and is unavailable.